P25 - updated automatic and direct frequency control to cleanup the code and allow P25 to better track frequency and PPM changes in the tuner and provide a cleaner implementation of the AFC component in the processing chain of all of the other decoders.
Changed C4FM symbol filter to reset frequency tracking values when the tuner's frequency or PPM values change.

Moved all frequency control pieces into the source.tuner.frequency package.

+++ src/source/recording/Recording.java
@@ -5,12 +5,12 @@ import java.util.ArrayList;
 import java.util.concurrent.RejectedExecutionException;
 
 import source.SourceException;
-import source.tuner.FrequencyChangeEvent;
-import source.tuner.FrequencyChangeListener;
 import source.tuner.TunerChannel;
 import source.tuner.TunerChannelProvider;
 import source.tuner.TunerChannelSource;
-import source.tuner.FrequencyChangeEvent.Attribute;
+import source.tuner.frequency.FrequencyChangeEvent;
+import source.tuner.frequency.FrequencyChangeListener;
+import source.tuner.frequency.FrequencyChangeEvent.Attribute;
 import controller.ResourceManager;
 import controller.ThreadPoolManager;
 
@@ -92,7 +92,7 @@ public class Recording implements Comparable<Recording>,
     {
 		if( event.getAttribute() == Attribute.FREQUENCY )
 		{
-			int frequency = (int)event.getValue();
+			long frequency = event.getValue().longValue();
 			
 			mConfiguration.setCenterFrequency( frequency );
